Fixed firefox button position issue
WARNING: DETAIL:
Due to the codemirror div being 100% of the container box, when the height of the fieldset containing the buttons is added to the size of the code, the Code Mirror box consumes that height leaving blank space below the code pushing the buttons below the div. This only happens on firefox as far as I can tell. Changing to height: intial (overriding the height: 100%) fixes this